requirements.txt
pytorch-lightning=0.5.3.2
pytorch==1.4.0
torchvision=0.4.2
pandas==0.24.2
numpy==1.16.4
scikit-learn==0.20.3